# Conflicts:
#	src/main/java/org/example/backend/models/Entity.java
#	src/main/java/org/example/backend/models/Item.java
#	src/main/java/org/example/backend/models/items/ArtItem.java
#	src/main/java/org/example/backend/models/items/ElectronicsItem.java
#	src/main/java/org/example/backend/models/items/VehicleItem.java

+++ src/main/java/org/example/backend/models/items/ArtItem.java
@@ -2,37 +2,38 @@ package org.example.backend.models.items;
 
 import org.example.backend.models.Item;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public class ArtItem extends Item {
-    protected String artist;
-    protected int creationYear;
+  protected String artist;
+  protected int creationYear;
 
-    public ArtItem(
-            int id,
-            LocalDateTime createdAt,
-            String type,
-            String itemName,
-            String artist,
-            int creationYear,
-            String description,
-            double startingPrice) {
-        super(id, createdAt, type, itemName, description, startingPrice);
-        this.artist = artist;
-        this.creationYear = creationYear;
-    }
+  public ArtItem(
+          int id,
+          LocalDateTime createdAt,
+          String type,
+          String itemName,
+          String artist,
+          int creationYear,
+          String description,
+          BigDecimal startingPrice) {
+    super(id, createdAt, type, itemName, description, startingPrice);
+    this.artist = artist;
+    this.creationYear = creationYear;
+  }
 
-    public String getArtist() {
-        return artist;
-    }
+  public String getArtist() {
+    return artist;
+  }
 
-    public int getCreationYear() {
-        return creationYear;
-    }
+  public int getCreationYear() {
+    return creationYear;
+  }
 
-    public void printInfo() {
-        System.out.println(id + ": " + getItemName());
-        System.out.println("Mô tả sản phẩm: " + description);
-        System.out.println("Giá khởi điểm: " + getStartingPrice());
-    }
+  public void printInfo() {
+    System.out.println(id + ": " + getItemName());
+    System.out.println("Mô tả sản phẩm: " + description);
+    System.out.println("Giá khởi điểm: " + getStartingPrice());
+  }
 }
